+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/AObjectTypePermission.cs
@@ -4,7 +4,7 @@ namespace IsusCoreFullNet2017Spa.IsusModels
 {
     public partial class AObjectTypePermission : Entity
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public int ObjectTypeId { get; set; }
         public int Permissions { get; set; }
