+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/FEducationResult.cs
@@ -5,14 +5,14 @@ namespace IsusCoreFullNet2017Spa.IsusModels
 {
     public partial class FEducationResult : Entity<long>
     {
-        public int PupilId { get; set; }
+        public long PupilId { get; set; }
         public int ActivityId { get; set; }
         public int PeriodId { get; set; }
         public DateTime DtStart { get; set; }
         public DateTime DtEnd { get; set; }
         public int Hours { get; set; }
         public int MarkMain { get; set; }
-        public int? PlanId { get; set; }
+        public long? PlanId { get; set; }
 
         public SActivity Activity { get; set; }
         public CPeriodType Period { get; set; }
