+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/FExpertStudyProgram.cs
@@ -4,7 +4,7 @@ namespace IsusCoreFullNet2017Spa.IsusModels
 {
     public partial class FExpertStudyProgram : Entity<long>
     {
-        public int ExpertId { get; set; }
+        public long ExpertId { get; set; }
         public int ActivityId { get; set; }
         public int Grade { get; set; }
         public bool? IsActual { get; set; }
