+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/FStudyJob.cs
@@ -11,8 +11,8 @@ namespace IsusCoreFullNet2017Spa.IsusModels
         public double Hours { get; set; }
         public int? BasicGroupId { get; set; }
         public int? ExtraGroupId { get; set; }
-        public int TeacherId { get; set; }
-        public int MetodistId { get; set; }
+        public long TeacherId { get; set; }
+        public long MetodistId { get; set; }
 
         public SActivity Activity { get; set; }
         public FGroupExtraItem Extragroup { get; set; }
