+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/FWorker.cs
@@ -5,7 +5,7 @@ namespace IsusCoreFullNet2017Spa.IsusModels
 {
     public partial class FWorker : Entity<long>
     {
-        public int WorkerId { get; set; }
+        public long WorkerId { get; set; }
         public DateTime DtStart { get; set; }
         public DateTime? DtEnd { get; set; }
         public int PositionId { get; set; }
