+++ IsusCoreFullNet2017Spa/3.0.0/aspnet-core/src/IsusCoreFullNet2017Spa.Core/IsusModels/UserParentship.cs
@@ -4,8 +4,8 @@ namespace IsusCoreFullNet2017Spa.IsusModels
 {
     public partial class UserParentship : Entity
     {
-        public int ParentId { get; set; }
-        public int ChildId { get; set; }
+        public long ParentId { get; set; }
+        public long ChildId { get; set; }
         public string RoleName { get; set; }
         public string RoleDescr { get; set; }
